Changes to Margin of Victory Calculation
Rationale: Favored teams tend to win by more, so the difference in ELO must be accounted for in the K_mult variable to not introduce autocorrelation. The idea with an ELO system is that over time a properly rated team will remain at the exact same level, no matter who they play because it is all based off of win-shares probability. But, when we introduce margin of victory, winning teams get more points because they win more often and by more. So, there must be some discount to the extra points awarded for margin of victory. See, https://fivethirtyeight.com/features/how-we-calculate-nba-elo-ratings/ and https://fivethirtyeight.com/features/introducing-nfl-elo-ratings/ for more information